Optimize cosmosdb output
Cosmosdb output was way to slow. It will now trigger 3 seperate orchestrations calls for outputting data.
