+++ delete_cosmosdb_items/function.json
@@ -2,8 +2,8 @@
   "scriptFile": "__init__.py",
   "bindings": [
     {
-      "name": "context",
-      "type": "orchestrationTrigger",
+      "name": "payload",
+      "type": "activityTrigger",
       "direction": "in"
     }
   ]
