+++ stocktracker_orchestrator/__init__.py
@@ -72,8 +72,22 @@ def orchestrator_function(context: df.DurableOrchestrationContext):
         "calculate_totals", [data, invested, transactions, days_to_update]
     )
 
-    # step 8 - Output to cosmos db
-    result = yield context.call_activity("output_to_cosmosdb", [data, days_to_update])
+    # step 8 recreate containers / remove items
+    result = yield context.call_activity("delete_cosmosdb_items", days_to_update)
+
+    # step 9.1 - output single_day_data to cosmosdb
+    result = yield context.call_activity("output_singleday_to_cosmosdb", data)
+
+    # step 9.2 - output everything else to cosmosdb
+    provisioning_tasks = []
+    id_ = 0
+    for container_name, items in data.items():
+        child_id = f"{context.instance_id}:{id_}"
+        provision_task = context.call_sub_orchestrator(
+            "output_to_cosmosdb", [container_name, items], child_id
+        )
+        provisioning_tasks.append(provision_task)
+    result = yield context.task_all(provisioning_tasks)
 
     return result
